+++ api/all/src/main/java/io/opentelemetry/api/metrics/DoubleGaugeBuilder.java
@@ -42,5 +42,5 @@ public interface DoubleGaugeBuilder {
    *
    * @param callback A state-capturing callback used to observe values on-demand.
    */
-  void buildWithCallback(Consumer<ObservableDoubleMeasurement> callback);
+  ObservableDoubleGauge buildWithCallback(Consumer<ObservableDoubleMeasurement> callback);
 }