+++ api/all/src/main/java/io/opentelemetry/api/metrics/ObservableLongMeasurement.java
@@ -8,29 +8,7 @@ package io.opentelemetry.api.metrics;
 import io.opentelemetry.api.common.Attributes;
 
 /** An interface for observing measurements with {@code long} values. */
-public interface ObservableLongMeasurement extends ObservableMeasurement {
-  /**
-   * Records a measurement.
-   *
-   * @param value The measurement amount.
-   * @deprecated Use {@link #record(long)}.
-   */
-  @Deprecated
-  default void observe(long value) {
-    record(value);
-  }
-
-  /**
-   * Records a measurement with a set of attributes.
-   *
-   * @param value The measurement amount.
-   * @param attributes A set of attributes to associate with the count.
-   * @deprecated Use {@link #record(long, Attributes)}.
-   */
-  @Deprecated
-  default void observe(long value, Attributes attributes) {
-    record(value, attributes);
-  }
+public interface ObservableLongMeasurement {
 
   /**
    * Records a measurement.