+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/state/AsynchronousMetricStorage.java
@@ -38,8 +38,9 @@ import java.util.logging.Logger;
  * at any time.
  */
 public final class AsynchronousMetricStorage<T> implements MetricStorage {
-  private static final ThrottlingLogger logger =
-      new ThrottlingLogger(Logger.getLogger(DeltaMetricStorage.class.getName()));
+  private static final Logger logger = Logger.getLogger(AsynchronousMetricStorage.class.getName());
+
+  private final ThrottlingLogger throttlingLogger = new ThrottlingLogger(logger);
   private final MetricDescriptor metricDescriptor;
   private final ReentrantLock collectLock = new ReentrantLock();
   private final AsyncAccumulator<T> asyncAccumulator;
@@ -56,15 +57,15 @@ public final class AsynchronousMetricStorage<T> implements MetricStorage {
       View view,
       InstrumentDescriptor instrument,
       Consumer<ObservableDoubleMeasurement> metricUpdater) {
-    final MetricDescriptor metricDescriptor = MetricDescriptor.create(view, instrument);
+    MetricDescriptor metricDescriptor = MetricDescriptor.create(view, instrument);
     Aggregator<T> aggregator =
         view.getAggregation().createAggregator(instrument, ExemplarFilter.neverSample());
 
-    final AsyncAccumulator<T> measurementAccumulator = new AsyncAccumulator<>(instrument);
+    AsyncAccumulator<T> measurementAccumulator = new AsyncAccumulator<>(instrument);
     if (Aggregator.drop() == aggregator) {
       return empty();
     }
-    final AttributesProcessor attributesProcessor = view.getAttributesProcessor();
+    AttributesProcessor attributesProcessor = view.getAttributesProcessor();
     // TODO: Find a way to grab the measurement JUST ONCE for all async metrics.
     final ObservableDoubleMeasurement result =
         new ObservableDoubleMeasurement() {
@@ -92,13 +93,16 @@ public final class AsynchronousMetricStorage<T> implements MetricStorage {
       View view,
       InstrumentDescriptor instrument,
       Consumer<ObservableLongMeasurement> metricUpdater) {
-    final MetricDescriptor metricDescriptor = MetricDescriptor.create(view, instrument);
+    MetricDescriptor metricDescriptor = MetricDescriptor.create(view, instrument);
     Aggregator<T> aggregator =
         view.getAggregation().createAggregator(instrument, ExemplarFilter.neverSample());
-    final AsyncAccumulator<T> measurementAccumulator = new AsyncAccumulator<>(instrument);
-    final AttributesProcessor attributesProcessor = view.getAttributesProcessor();
+    AsyncAccumulator<T> measurementAccumulator = new AsyncAccumulator<>(instrument);
+    if (Aggregator.drop() == aggregator) {
+      return empty();
+    }
+    AttributesProcessor attributesProcessor = view.getAttributesProcessor();
     // TODO: Find a way to grab the measurement JUST ONCE for all async metrics.
-    final ObservableLongMeasurement result =
+    ObservableLongMeasurement result =
         new ObservableLongMeasurement() {
 
           @Override
@@ -147,7 +151,7 @@ public final class AsynchronousMetricStorage<T> implements MetricStorage {
         metricUpdater.run();
       } catch (Throwable e) {
         propagateIfFatal(e);
-        logger.log(
+        throttlingLogger.log(
             Level.WARNING,
             "An exception occurred invoking callback for instrument "
                 + getMetricDescriptor().getName()
@@ -175,7 +179,9 @@ public final class AsynchronousMetricStorage<T> implements MetricStorage {
   }
 
   /** Helper class to record async measurements on demand. */
-  private static final class AsyncAccumulator<T> {
+  // Visible for testing
+  static final class AsyncAccumulator<T> {
+    private final ThrottlingLogger throttlingLogger = new ThrottlingLogger(logger);
     private final InstrumentDescriptor instrument;
     private Map<Attributes, T> currentAccumulation = new HashMap<>();
 
@@ -184,8 +190,9 @@ public final class AsynchronousMetricStorage<T> implements MetricStorage {
     }
 
     public void record(Attributes attributes, T accumulation) {
+      // Check we're under the max allowed accumulations
       if (currentAccumulation.size() >= MetricStorageUtils.MAX_ACCUMULATIONS) {
-        logger.log(
+        throttlingLogger.log(
             Level.WARNING,
             "Instrument "
                 + instrument.getName()
@@ -194,7 +201,17 @@ public final class AsynchronousMetricStorage<T> implements MetricStorage {
                 + ").");
         return;
       }
-      // TODO: error on metric overwrites
+
+      // Check there is not already a recording for the attributes
+      if (currentAccumulation.containsKey(attributes)) {
+        throttlingLogger.log(
+            Level.WARNING,
+            "Instrument "
+                + instrument.getName()
+                + " has recorded multiple values for the same attributes.");
+        return;
+      }
+
       currentAccumulation.put(attributes, accumulation);
     }